django/requirements.txt
appdirs==1.4.3
asgiref==3.2.7
CacheControl==0.12.6
certifi==2020.4.5.1
chardet==3.0.4
cliapp==1.0.9
click==7.1.1
colorama==0.4.3
contextlib2==0.6.0
crayons==0.3.0
distlib==0.3.0
distro==1.4.0
Django==3.0.5
django-appconf==1.0.4
django-cors-headers==3.2.1
django-countries==6.1.2
django-friendship==1.9.1
django-ipware==2.1.0
djangorestframework==3.11.0
html5lib==1.0.1
idna==2.9
ipaddr==2.2.0
jsonfield==3.1.0
lockfile==0.12.2
msgpack==0.6.2
numpy==1.18.3
packaging==20.3
pep517==0.8.2
Pillow==7.1.1
pinax-badges==2.0.3
pinax-cli==1.1.4
pinax-points==1.0
pinax-stripe==4.4.0
progress==1.5
pyparsing==2.4.6
pytoml==0.1.21
pytz==2019.3
requests==2.23.0
retrying==1.3.3
six==1.14.0
sqlparse==0.3.1
stripe==2.45.0
urllib3==1.25.9
webencodings==0.5.1
